+++ src/framework/index.css
@@ -30,7 +30,7 @@
 @import url("../lib/components/navigation/anchor/anchor.theme.css");
 @import url("../lib/components/navigation/breadcrumb/breadcrumb.theme.css");
 @import url("../lib/components/overlays/clickable/clickable.theme.css");
-@import url("../lib/components/pickers/picker/picker.theme.css");
+@import url("../lib/components/widgets/widget/widget.theme.css");
 @import url("../lib/components/surfaces/box/box.theme.css");
 @import url("../lib/components/surfaces/card/card.theme.css");
 @import url("../lib/components/surfaces/hero/hero.theme.css");
@@ -77,7 +77,7 @@
 @import url("../lib/components/navigation/menu/menu.css");
 @import url("../lib/components/navigation/omni/omni.css");
 
-@import url("../lib/components/pickers/picker/picker.css");
+@import url("../lib/components/widgets/widget/widget.css");
 
 @import url("../lib/components/surfaces/card/card.css");
 @import url("../lib/components/surfaces/box/box.css");
