--- src/lib/components/pickers/picker/index.ts
@@ -1,5 +1,0 @@
-export {default as Container} from "./PickerContainer.svelte";
-
-export {default as Button} from "./PickerButton.svelte";
-export {default as Header} from "./PickerHeader.svelte";
-export {default as Section} from "./PickerSection.svelte";
