+++ src/lib/components/widgets/monthpicker/MonthPicker.stories.svelte
