+++ src/lib/components/widgets/widget/WidgetSection.svelte
