+++ src/lib/components/widgets/widget/index.ts
@@ -1,0 +1,5 @@
+export {default as Container} from "./WidgetContainer.svelte";
+
+export {default as Button} from "./WidgetButton.svelte";
+export {default as Header} from "./WidgetHeader.svelte";
+export {default as Section} from "./WidgetSection.svelte";
